backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# ML/Embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
torch==2.1.0

# Utils
python-dotenv==1.0.0
gitpython==3.1.40
aiofiles==23.2.1

# Optional (for Gemini)
google-generativeai==0.3.1
